refactor(Accordion): move translation into parent
