+++ app/components/Accordion.tsx
@@ -1,5 +1,9 @@
 import { useRef } from "react";
-import AccordionItem, { type AccordionItemProps } from "./AccordionItem";
+import AccordionItem, {
+  type AccordionItemProps,
+} from "~/components/AccordionItem";
+import { getTranslationByKey } from "~/services/translations/getTranslationByKey";
+import { useTranslations } from "~/services/translations/translationsContext";
 
 export type AccordionProps = Readonly<{
   items: AccordionItemProps[];
@@ -7,6 +11,11 @@ export type AccordionProps = Readonly<{
 
 export default function Accordion({ items }: AccordionProps) {
   const itemsRef = useRef<HTMLDetailsElement[]>([]);
+  const { accordion } = useTranslations();
+  const labels = {
+    show: getTranslationByKey("accordionItemShow", accordion),
+    hide: getTranslationByKey("accordionItemHide", accordion),
+  };
 
   return (
     <section className="border-2 rounded-lg border-blue-500">
@@ -15,6 +24,7 @@ export default function Accordion({ items }: AccordionProps) {
           key={item.title}
           title={item.title}
           description={item.description}
+          labels={labels}
           ref={(el) => {
             if (el) itemsRef.current[index] = el;
           }}
